Make opaque and alpha-masked effects show up in Bevy 0.16.
PR #445 fixed the problem for transparent effects, but opaque and
alpha-masked effects still didn't show up. The problem here was the use
of `MainEntity::PLACEHOLDER` when inserting the custom phase items into
a bin. As part of the new retained bins feature, Bevy 0.16 clears out
items in bins that don't correspond to visible entities every frame. So
it was deleting every binned phase item that Bevy added.

This patch fixes the issue by tracking the main entity that each effect
corresponds to and placing that in the phase item. Additionally, I made
opaque effects actually go into the opaque bin instead of the alpha mask
bin; I assume this was the intended behavior.