Mark `CachedParentInfo` as changed if any `CachedChildInfo`did (#451)
Whenever a `CachedChildInfo` is created, then it must be "fixed up" via
the `fixup_parents` system. That system uses a
`Changed<CachedParentInfo>` query filter to determine which parents it
needs to process the children of. Usually, `resolve_parents` will
regenerate the `CachedParentInfo` of every parent with a changed child,
because the `is_child_list_changed` check will notice that the child
entity IDs differ, and thus `fixup_parents` will pick up that
newly-created `CachedParentInfo`. However, `resolve_parents` will skip
regenerating the `CachedParentInfo` if the entity IDs of the children
didn't change (as determined by `is_child_list_changed`). This means
that, if the children changed but for whatever reason their *entity IDs*
didn't, the change flag will never be set on the parent's
`CachedParentInfo`. This means that `fixup_parents` will never process
the `CachedChildInfo`, causing a crash later as the `global_child_index`
will be wrong.

This PR fixes the problem by modifying `resolve_parents` so that,
whenever a new `CachedChildInfo` is built, the parent of that child is
added to a set. Then `resolve_parents` forcibly marks the
`CachedParentInfo` of any parent in that set as changed, even if
`is_child_list_changed` determined that the *contents* of the
`CachedParentInfo` didn't need to change. This ensures that
`fixup_parents` examines those parents and updates the `CachedChildInfo`
of their children, preventing the crash.