Allow multiple effects to be packed into a single buffer again.
This commit restores the functionality present in older versions of
Hanabi that allows multiple effects to be packed into a single buffer.
This is in preparation for reenabling batching.

I use the `base_instance` field on the effects metadata to supply the
offset within each indirect buffer to the shader. This is convenient,
because it naturally paves the way to use multi-draw as a follow-up, and
it avoids having to add more fields to the effects metadata.

As suggested in the code, I rewrote sorting to be prior to batching
instead of after batching. It now uses a modified version of [Tarjan's
algorithm] to sort the effects into the optimum batching order. Not only
does it make sure that parents precede children, but it also groups
buffers together, and sorts by offset, in order to make batching easier
as a follow-up. Sorting by offset within the buffer should enable the
use of binary search in order for the shader to map from the global
thread ID to the effect instance.

[Tarjan's algorithm]: https://en.wikipedia.org/wiki/Topological_sorting#Depth-first_search