Fix sorting when multiple effects are packed into a single buffer.
The sort shaders weren't updated, which caused failures in the `ribbons`
example. It looks like the plan was to use dynamic offsets in order to
move the window around within the buffer; however, that doesn't work due
to alignment limitations. So I removed the dynamic offsets and instead
use the new field in the effects metadata.

Much of the complexity comes from the fact that the render shader now
needs access to the effects metadata (in just a single place), because
there's no special WGSL shader variable available for the base instance.
Thus I had to update a bunch of bind groups and such in order to make
the effects metadata accessible to the render shader. The good news is
that, if we ever need to add more fields to the effects metadata that
the render shader needs to access in the future, doing so will be
straightforward.