Make sure `GpuEffectMetadata` gets initialized even if the mesh didn't change on the frame the effect was first extracted.
It's possible for an effect to have a slot in the `GpuEffectMetadata`
buffer reserved, but remaining uninitialized, if that effect has a
`Visibility::Hidden` component on the first frame it was spawned. In
such cases, Hanabi currently never initializes the `GpuEffectMetadata`
at all, because it gates initialization of the metadata behind the mesh
changing, which it won't in this case.

This commit fixes the bug by recording which effects have
`GpuEffectMetadata` entries in a `MainEntityHashSet` in `EffectsMeta`
and, if the effect doesn't have an entry in that set, creates a
`GpuEffectMetadata` entry for it anyway, even if the mesh didn't change.