fix(examples/billboard): filter out ui camera  (#496)
In the billboard example, the automatic camera rotation doesn't work as
it filters for any cameras, and trying to get a single mutable camera,
which won't work as `utils::DemoApp` also spawns a Camera2d for the UI.
This breaks `single_mut` as now there are two cameras to query for.

This can easily be fixed by filtering for a `Camera3d`. Also, I've used
the Single query filter to get rid of one level of indentation (: