Pack multiple effects per particle slab (#508)
Sub-allocate each particle slab with the content of multiple effects.
- Introduce a `base_particle` value per effect instance, which is the
equivalent of the `base_vertex` for rendering, and corresponds to the
index of the first particle in the sub-allocated slice for that effect,
inside the overall slab buffer.
- Store that `base_particle` in the `Spawner` and give access to all
shaders which need it (most of them).
- Restore the default 64k particle count per slab, which allows packing
multiple effects per buffer/slab.

When `debug_assertions` is active (in Debug build), fill the first value
of each particle to a `NaN` (0xFFFFFFFF) to make it easier to see in
RenderDoc or any other GPU debugger that the particle is unused.